added signs for finish and y1
in signs.py we added signs for finish and Y1, and we took care of what happens
if the result is only a constant

created new file that calculates the difference in signs from the isotopy
that takes a straight flow line and returns a tree with an end and a Y1

saved the results in relations_from_eY1isotopy. Not everything is consistent
with the calculations in the long article, must check.

Do:
Recalculate the intersection signs for Y1end-isotopy
Consider the special case when n=1