created python file to compute the different signs if we introduce a Y1-end constellation in a tree with two Y0-vertices.
Have coded one of the easiest case, when the Y1-end is introduced just before the
positive puncture
Some mistake, since the difference between the stabilization signs
of the two different trees we might have differs by 1 compared to the pair of
trees with only Y0:s. It should give the same difference.